[roll] Roll fuchsia [zircon][fidl][zx] Re^5land: Extract built-in zx library
This moves the magically injected `library zx` out of fidlc, and into
normal fidl files in //zircon/vdso. This allows the syscall definitions
to share the definition with other fidl, and also to have the rest of
the syscall definitions live in library zx (rather than the fake
"library zz" that was internally mapped to zx by kazoo).

Changes from previous attempts:
- Dependency on zx is explicit in all fidl libraries
- Remove [Internal] from zx
- fostr automatically depends on the zx fostr
- gidl ignores a zx dependency (as library zx is no longer Internal.)

Original-Bug: 7802
Original-Bug: 42717
Original-Bug: 44602
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380974
GitOrigin-RevId: c9b527764bf8c3b534a91f24dc51ac80f1828f90
Change-Id: I94ba305ac0e0e1158ea8268d1c0d7b966fb4218c