[roll] Roll fuchsia [debugger] Add location support for clear/enable/disable
Introducing a new syntax for clear/enable/disable which takes an
optional location so that the behavior is similar to GDB.

If a location is given, these commands will ignore current active
breakpoint and modify all breakpoints at that location.

If no location is given, the behavior should stay unchanged.

Original-Bug: 5452

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383407
GitOrigin-RevId: d6fe4b53d81c176d7d050ba27c4f6919856700ee
Change-Id: Iae2dd749af538fb6fd3da01e414418902e6b6fcb