[roll] Roll fuchsia [runtests] Avoid infinite loop
When an empty VMO is published, we would continue the loop, but
since we only popped the element from the list at the end of the
loop, this would never happen and we would enter an infinite loop.
So instead, pop the element at the beginning of the loop. We also
deduplicate the VMO mapping logic between the initialization and
merge loop.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383543
GitOrigin-RevId: 9e2f4307a3a6472a1fddedac473080f91dce50e9
Change-Id: I30c5e063cab008646a8ad5ea92680f484b79d5cc