[roll] Roll fuchsia [tools][debug][elflib] Handle PE-COFF PDB GUIDs as build IDs
This makes elflib a bit of a misnomer, since it reads both ELF
and PE-COFF now.  It's really buildidlib anyway, more of its code
is about build IDs abstractly than about the ELF format.

PE-COFF debug files (PDB files) are in a disparate format that is
complicated to decode and doesn't have a handy Go library.  So
instead of decoding PDB files to match the build ID (GUID + age) in
the PE-COFF binary, a simple "is a PDB" detector is provided so that
buildidtool can do a kludge to suppress the build ID matching and
still link the .pdb file as the debug file.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381556
GitOrigin-RevId: 66e9aee4fab9b53220c83176d74397513d9a540f
Change-Id: Ie3ddb71179ff408bb2a7cb5b75f39913e0733958