[roll] Roll fuchsia [bt][hci] Avoid assuming kernel message batching behaviour in tests.
The current test "HCI_CommandChannelTest.EventHandlerResults" implicitly
assumes a particular behaviour of how the Zircon kernel will notify
userspace about batches of received messages in a channel. This
behaviour is changing, causing the test to start failing.

This CL moves the event callbacks in the test to use their own
loop/dispatcher, so that the test can independently control when the
event callbacks are processed (as is required for the test) instead of
this implicitly occurring at the end of each batch of messages on the
channel.

We additionally split out testing of "SendCommand" behaviour when an
event handler has already been registered into its own test.

Original-Bug: 50601

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381895
GitOrigin-RevId: eb84a181e80bd52e8097ff68b3b74c46230c5a6b
Change-Id: I5e85b03ff0206c5d2ae6eeff321a269f40c987e8