[roll] Roll fuchsia [storage] Add Zero method to storage::BlockBuffer.
This is needed for following changes. Where the buffer is backed by a
VMO we can use a specialised zero method. Also change the interface so
that it's block based like the rest of the methods.

Also renamed VnodeBufferType -> ResizeableBufferType since it's not
specific to a Vnode.

Removed the unit tests for VnodeBufferType since we should alrady have
coverage for the Zero method elsewhere.

Original-Bug: 42096
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382651
GitOrigin-RevId: 217ec74fcda4d47b8605d13abdd8e315131e6916
Change-Id: I75c034fcbb490454a2c429a2ce1c9de92e843037