[roll] Roll fuchsia [fidlc] Recoverable parsing at member boundary
This CL builds on the functionality added in fxrev.dev/378714 to add
recoverable parsing between members of a declaration. E.g., if a table
member is missing an ordinal, the parser will attempt to recovr to the
next member of the table and continue parsing, rather than skipping to
the end of the table.

In future CLs:
* Recovery within a parameter/attribute list
* Token insertion
* Recoverable error reporting in the compilation stage

Test: fidlc host tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381145
GitOrigin-RevId: ffe88ae901d9f015e196d3d1284d9fd8696bdb3d
Change-Id: Iee2cad2ceec2d618666ab79140b5700b63534f44