[roll] Roll fuchsia [escher] Define lifetime of DescriptorSetAllocators
This CL moves caching policy for DescriptorSetAllocators into
DescriptorSetAllocatorCache class. DescriptorSetAllocatorCache hands
out RefPtrs and does not extend the lifetime by holding WeakPtrs.

As a result, their lifetime is tied to PipelineLayout objects. Since
we are lacking a cleanup policy in PipelineLayoutCache, this CL adds
a HashCache to release them if they haven't been used.

Original-Bug: 7289
Test: escher_tests -t escher_unittests -v --
--gtest_filter=*DescriptorSetAllocatorCache*

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/373593
GitOrigin-RevId: f6fcdddcf32b35eb54feffa21bdbc8169f1388db
Change-Id: Id575ffec73aa028d36e2b22157a92c26680c084e