[roll] Roll fuchsia [images] support firmware_prebuilts
Adds support for the firmware_prebuilts variable, which allows adding a
set of firmware images to the build package.

This is a replacement for the existing bootloader_prebuilt variable
with the added ability to specify multiple firmware images of different
types - for example, for multi-stage bootloaders provided in separate
files.

Original-Bug: 45606
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381817
GitOrigin-RevId: 183326df1a93210023a60ee8ebc3462882722807
Change-Id: I0a16fa8c6768b3a92f78cbff950e92ab12f0f7c9