[roll] Roll fuchsia [archivist] Don't fail on send if receiver is closed
The new API might abort the archivist future while the handling of
an event is going on on a separate task. When aborted the receiver end
of the channel is closed, this leads to SendError (can only occur when
the other end of the channel is missing). This caused a flake in logs.

Original-Bug: 50850
Tested: existing tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383875
GitOrigin-RevId: 3529727d0aba94f3131801a8bc2e7b2178d014f4
Change-Id: I235c98d3e298df54519c2849fcd4bfd91916a9f7