[roll] Roll fuchsia [cobalt] Make FuchsiaSystemClock thread safe.
A pointer to an instance of FuchsiaSystemClock is passed in
to the CobaltConfig that is passed to the CobaltService
in Cobalt core. Currently the now() method is only ever
invoked from the Logger::Log*() method and so it is only
ever invoked on the main thread of the Cobalt FIDL service.
However in the future I can imagine we might want to invoke
the now() method from one of Cobalt core's background threads.
This CL makes it safe to do so.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383354
GitOrigin-RevId: f6333a3be9d3a411d8d73d155de947cd6c21711a
Change-Id: I30d1cc3dc05841362f33134c200b123b1cb14dfb