[roll] Roll fuchsia [media][audio_core] Parse config driver_gain_db
Parse a new driver_gain_db field from the output device configuration
and include it in the device profile. In a following CL, the driver gain
value will be applied to the driver upon initialization.

This is the first in a chain of CLs that will allow driver gain to be
modified in AudioCore on a per-device basis and in relation to audio
effects. Enabling driver_gain setting in the device configuration
ensures the driver's default gain setting will be Unity (0 db) unless
modified in the configuration.

Driver gain modifications are meant to be temporary fixes for loudness
issues and/or to compensate for a lack of device tuning.

Original-Bug: 49302

Test: fx test audio_core_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383403
GitOrigin-RevId: f101ae2ada3cff101bc6e4f3ae3d9e433a0515f3
Change-Id: I01b3e88276da1a0cd2bbd0a5231a8bff9af15278