[roll] Roll fuchsia [logger] Support logging the source file and line number.
Specifically, this enables support for `log.Lshortfile` and
`log.Llongfile`. We needed to refactor to allow tracking the call
depth because otherwise the log would have always said it came from
the `Printf` line in `logger.go`.

Go `log` provides us with `Output` instead of `Printf`, which takes in a
call depth that tells it how far down the call stack it should go to
find the actual log line.

Original-Fixed: 50257
Test: logger_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381362
GitOrigin-RevId: 8f6c7b36fc4c8ed206c8f54707b2ba8aea847715
Change-Id: I31cc3c1c957ea42b3173a5a11b66ee46dcb68051