[roll] Roll fuchsia [mmio] Add low-level MMIO API for buffer accesses
This CL introduces the MMIO_PTR macro which can be used for decorating
pointers to indicate they point to mmio buffers. Along with it, are
functions for doing simple reads/writes to these MMIO pointers.

Drivers should use this for accessing mmio buffers. This will eventually
replace hw/reg.h in the ddk and reg.h in the kernel.

We should check whether `uint8_t MMIO_PTR*` or `MMIO_PTR uint8_t*` will
be the standard style in Fuchsia.

The next step after this lands is to see which drivers will break from
this and target a select few to be migrated. Afterwards we can send out
a PSA letting driver writers know that this should be the API used for
accessing MMIO buffers.

Original-Bug: 32730
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380707
GitOrigin-RevId: 69ef259925788ab362fa1d763f6c3df644a6fbc7
Change-Id: I808cfe77e7112e4422dabd654ad2af4d498a9942