[roll] Roll fuchsia [fidlgen][libfuzzer] Generate goldens for libfuzzer
This CL fixes a bug in regen.sh which prevented it from generating
libfuzzer goldens. It is only supposed to emit them for FIDL libaries
that have a nonempty protocol; the bug was that it tested for this
assuming $src_path was a FIDL file, when it is actually JSON IR.

This CL also fixes some inaccuracies in the contributing_to_fidl.md
goldens table.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383906
GitOrigin-RevId: 12fa65d1db77ab8c3c530600daa579e273fa3bca
Change-Id: I9cedf4e796f964d428e77b0d9ec4546a41ebda70