[roll] Roll fuchsia [kernel][unittest] Clean up testing::UserMemory API and use
Most uses were repeating the work of ->user_in and ->user_out locally.
Instead, just use them.  Replace the ->in and ->out methods for the
few direct accesses with ->get and ->put that bottleneck all such
accesses.  This ensures that there is a single place to special-case
for memory access instrumentation such as ASan that needs to treat
these accesses differently than access to normal kernel memory.

Original-Bug: 30033
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383642
GitOrigin-RevId: 954eb88efcd24b28455beb9a5588137b1ad9236a
Change-Id: I87d16475369b1e092ea6f9d58c0ea94032676a0b