[roll] Roll fuchsia [storage] Factor out FDIO test harness for blobfs
Factors out the current test setup from the mount tests into a separate
file. This will allow the non-trivial fdio/blobfs setup to be shared
with other tests.

The remote ramdisk is removed from this test. Instead, it maps to a fake
block device which allows for simpler test setup and will allow more
tests to be done on blobfs behavior in the future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383926
GitOrigin-RevId: 4bef92f40effc15877504b7ad2f6de9e9737675f
Change-Id: I620238ee3247eccce7cdc6c8f2f14df411c017d5