[roll] Roll fuchsia [fidlgen] Exclude declarations based on BindingsDenylist
This CL makes fidlgen backends omit FIDL declarations based on the
BindingsDenylist attribute. This is meant to be used in conjunction with
bindings_denylist in GIDL.

For example:

    [BindingsDenylist = "rust,go"]
    struct Foo {};

For this FIDL file, the struct Foo would be absent from the Rust and Go
bindings, as if it was never written. The attribute supports all
backends except dart: cpp, llcpp, rust, go, dart, libfuzzer, syzkaller.
A follow-up topaz CL will add support in fidlgen_dart.

Test: goldens
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383341
GitOrigin-RevId: 70f34e56da0f9223a4fbd481204aba12b38bd389
Change-Id: I936254370e28b4014dbc0cff4e4b048aa89073a4