[roll] Roll fuchsia [syscalls][kazoo] Fix a few Go remapping errors
There were a few types that didn't have appropriate name hacks added for
them. Kazoo errors out on attempting to run regeneration, but because Go
isn't run as part of the build (it's in a separate repo) they weren't
noticed.

As a way of avoiding this in future, generate all possible output styles
to a golden file, and diff in a build rule to catch failures in all
output styles, as well as potentially unintentional changes.

This has the possibly unfortunate side-effect of catching simple
intentional changes when, e.g. a syscalls arguments change, but it's
probably a reasonable idea to have an second intentional review of the
output generated for those changes anyway.

Test: zircon/tools/kazoo/golden-test added
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379092
GitOrigin-RevId: 1bd9ca3cb9538d005c8e5409efa9edba2bef11e0
Change-Id: I1deaf33f0dbc8fc8136c9834d8cebd20f5fd547c