[roll] Roll fuchsia [netstack_iperf_test] Use long command line option names for better readability
The single-char option names are less readable.

The change is based on the docs at third_party/iperf/src/iperf3.1.
(In contrast, the page https://iperf.fr/iperf-doc.php differs on the
longer names for "-b" and "-n".)

Update a comment based on this part of the docs:

       -n, --bytes n[KM]
              number of bytes to transmit (instead of -t)

Original-Bug: none
Test: on CQ: terminal-x64-release
 * run locally with: fx run-e2e-tests netstack_iperf_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383334
GitOrigin-RevId: 52c13fb5557698a6ba26637e765c9ca2f9a435f1
Change-Id: Idf146e1ef1c361f0a692364c6b41771d84b82767