[roll] Roll fuchsia [audio][tests] Make ACCESS_DENIED benign unless --admin specified
The audio_driver_tests binary directly connects to audio drivers. A
subset of its test cases (16/36) can execute normally even when the
audio_core service is already running and connected to the driver.
Previously, the remaining 20/36 would fail with repeated messages of
ZX_ACCESS_DENIED error, if audio_core was connected to the driver.

Now, these tests will attempt to run, but if ZX_ACCESS_DENIED occurs,
this is considered a benign failure and the test cases are skipped.
A new flag "--admin" specifies that ACCESS_DENIED should be considered
a failure (generally because the tests are being run in an environment
that guarantees that no other party will attempt to connect to the
audio driver). Additionally, after ACCESS_DENIED is encountered,
subsequent test cases will automatically skip/fail, avoiding
additional log spam. Specifically, after this change, the
"Unprivileged channel cannot SetStreamFormat" error message should be
logged only twice (one for output driver, one for input driver) across
the 20 test case skips or failures.

Also, on targets with multiple audio input or output devices, the test
connects to all devices and stores a vector of the channels. For now
however, we still test only one (the last, previously tested only the
first). Groundwork is in place to test multiple devices in the future,
if needed.

Finally, gtest Parameterized tests are now used, to eliminate code
duplication across the testing of input and output devices. During
this integration, the suite was refactored into admin tests, basic
tests and a shared test_base class.

Test: test-only code, no production change
Original-Bug: 50109

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380715
GitOrigin-RevId: 7834ae14561ef64ac8f7b84a03c9753982581103
Change-Id: I017a27d2204f8b327bac458512c3ee03673c9bac