[roll] Roll fuchsia [flatland] Buffer registration and Image creation.
BufferCollections back Images and other content. Successful
registration of a buffer collection is effectively a successful call
to the flatland::Renderer API of the same name: Flatland caches a
mapping from the user-generated ID to the renderer-generated ID and
does nothing else with the BufferCollection.

Image creation references a registered BufferCollection. In order to
successfully create an Image, the BufferCollection must be allocated,
meaning all clients holding a Sysmem token must set their
constraints. Flatland blocks a short time waiting for buffer
allocation and caches metadata about the collection if it is
allocated.

This CL only includes creation/registration functions for Buffers and
Images. Releasing Buffers and Images and attaching Images to
Transforms will be in follow-up changes.

TEST=fx test -o flatland_unittests

MULTIPLY: `[
  {
    "name": "flatland_unittests",
    "os": "fuchsia",
    "total_runs": 40
  }
]`

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381414
GitOrigin-RevId: 618755e2eddcc56368a8d01dc0de8d5fd8416f6f
Change-Id: I1f341fa419d0cbed4c6892ce952cd04023c6e69b