[roll] Roll fuchsia [appmgr][test] Fix flake in component events integration test
This test was completely wrong. It was connecting to
Componenteventprovider on an environment but not coming from a
component. Hence the connection was failing as the component
NO_COMPONENT is not allow-listed to connect to ComponentEventProvider.
This was the cause of the flake. As a fix, the test was replicated but
using a real component.

Tested: component_events_integration_test
Original-Fixed: 49600

MULTIPLY: component_events_integration_tests: 100

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384136
GitOrigin-RevId: 0b115540b39981f3fd756d8579f4da7881b55b31
Change-Id: I4628d2b9aeff167d8b3c9bc152a59bb2c7020a03