[roll] Roll fuchsia [syscalls] Add zx_msi_allocate and zx_msi_create
This CL introduces two new syscalls that expose Message-Signaled
Interrupts (MSIs) to userspace.
zx_msi_allocate allocates a block of MSIs using MsiAllocation and is
gated by the root resource. MsiAllocationDispatcher has been added to
provide an interface to this object. zx_msi_create is used to create
an interrupt object corresponding to a given MSI id within an
MsiAllocation and requires a handle to a valid allocation.  The
userspace interface to an MSI is a derived InterruptDispatcher, so
existing interrupt syscalls work.

Documentation, script updates, and test have been added corresponding
to these new syscalls. The new object types and syscalls are tested in
both zircon kernel unit tests and zircon coretests.

Original-Bug: 32978

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375078
GitOrigin-RevId: 5ff76b0264628607e876d1c82204647ddf38e278
Change-Id: I2d246a1aa62c03ed508b6cb73f747a7b993baf6a