[roll] Roll fuchsia [build] Add kasan variant
This adds the "kasan" variant for asan in the kernel, so e.g. a
command like `fx set bringup.x64 --auto-dir --variant kasan` will
set up a build with ASan enabled in the kernel but not userland.
When kernel asan is fully implemented, this will merge back into the
"asan" variant.  This just adds the basic build support and runtime
stubs to make the kernel link.

Original-Bug: 30033
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383323
GitOrigin-RevId: a271aafcce1c78ac506030ab4962a537c3483dbb
Change-Id: I384b767c0055913d48b6bf058698ead976810d81