[roll] Roll fuchsia [sl4f][catapult] Enable uploading to Catapult for tests converted to SL4F
Enable uploading to Catapult for the SL4F versions of performance
tests in fuchsia.git.  We can do that now that the main changes from
issue 49592 have landed: the results will be uploaded under the name
"fuchsia-x64-nuc" rather than "terminal-x64-release", so we won't have
to break continuity in the future if we change which builder produces
these results.

Remove the uploadToCatapultDashboard argument; we no longer need a
per-test flag for disabling this.

Original-Bug: 23091
Test: on CQ: terminal-x64-release
 * This is difficult to test fully before committing.  I will check
   after this lands that results actually get uploaded to Catapult.  I
   have already checked that the terminal-x64-release CI builders get
   the correct env vars set.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383613
GitOrigin-RevId: bebac30642f0e813857dab5ddcc80378d2fa8287
Change-Id: I7af72986198fe2a0b6bd97eccc7e5d015f0cca51