[roll] Roll fuchsia [x86][pci] Introduce PCI RootHost
PciRootHost replaces the static allocators used in the
x86 board file with an interface that can be implemented
on different platforms for PCI Root support. Additionally,
Pciroot is now templated on the type of platform context necessary,
so bringing up PCI on a non-ACPI system should be similar to setting
up the basic board file as one would for a normal driver. For any
architecture to support PCI it will now only need to provide a
Pciroot implementation for that context and instantiate a RootHost.

Some other changes were made to make the resource logic in the ACPI
parsing more clear to folks who don't speak ACPICA.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/333310
GitOrigin-RevId: dc99aaa76e131d8eb30c59e4f0e9541ff0648334
Change-Id: Id2591163a6490cf81af5fee84c2ca18034189f7c