[roll] Roll fuchsia [scenic] Remove debug_name from scene graph dump.
The scene graph dump is one of the most useful tools for debugging
problems that occur in the field.  We want to include it in feedback
reports, so we remove the debug_name, which might potentially contain
PII.

Instead we use two new Inspect fields to expose the mapping between
View/ViewHolder ID <--> debug_name; these fields will be filtered out
of feedback reports.

Original-Bug: 48462

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383360
GitOrigin-RevId: b95eee3c394ca1adbb495b046a340f8912264bf6
Change-Id: I526263a60b849df54e036edce2ae5575d1166128