[roll] Roll fuchsia [fbl][containers] Reorder Containable template params.
Reorder the template parameters used by SinglyLinkedListable,
DoublyLinkedListable, and WAVLTreeContainable, to make them a bit
easier to use.

Usually, most users just have to specify the pointer type to use with
their mix-in class.  There are now two different types of options
which users may need to specify if they want to use more advanced
features.  The first is a TagType which can be used with the
ContainableBaseClasses helper to make it a bit easier to allow a type
to exist in multiple different containers at the same time.  The
second is a NodeOptions value which gives users the ability to
customize the behavior of their nodes, such as allowing or disallowing
copying of the node at compile time.

For the most part, use of these features is currently pretty limited
(which is part of the reason why now is the time to make this change).
When users do use one of these feature, they tend to use only one of
them and not the other.  This leaves a situation where one of those
cases is going to force the user to explicitly provide the default
template option in order to specify the non-default option they wanted
to control.  Since NodeOptions was the most recent addition, this
means that someone who wanted custom node options had to say something
like:

  fbl::DoublyLinkedListable<Obj*, fbl::DefaultObjectTag,
                            fbl::NodeOptions::AllowCopy>

So, to make this a bit easier:
1) Reorder the options from <PtrType, Tag, Options> to
   <PtrType, Options, Tag>
2) Introduce a Tagged using alias which reverses the order, putting
   the emphasis on the tag instead of the option.

So, now, users can say things like

class MyClass : public fbl::DoublyLinkedListable<MyClass*, MyOptions>,
                public fbl::ContainableBaseClasses<
                  fbl::TaggedDoublyLinkedListable<MyClass*, Tag1>,
                  fbl::TaggedWAVLTreeConatinable<MyClass*, Tag2>,
        /* OK */  fbl::TaggedSinglyLinkedLisable<MyClass*, Tag3, SLLOpt>,
   /* Also OK */  fbl::SinglyLinkedLisable<MyClass*, SLLOpt, Tag4>
                > { ... };

Finally, as part of this, get rid of the various
kDefault{ContainerType}NodeOptions constants which were being defined.
They don't really serve any good purpose anymore now that users can
get the default from using either the Tagged or not-Tagged version of
the mix-in.

Tests: updated existing unit tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382691
GitOrigin-RevId: f56d2f24a63611001b352cd5425b1a2209ccac0e
Change-Id: Iaa26248b4bb69353d6815b5aa0f467aba497fac2