[roll] Roll fuchsia [fidl][benchmarks] Add benchmarks for unions
This is a starter set of union benchmarks, more might be needed, e.g.:
- A struct with many unions (if this is too small to highlight union
cost)

I don't think unions need benchmarks for reserved fields like tables
because setting fields in larger unions should be about the same
performance as doing so in smaller unions. If this is true, then
reserved/not reserved shouldn't matter. If it isn't true, then it should
be made true.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382015
GitOrigin-RevId: efc55c0b0c85e7029bb7297e5dd26ca5012266bb
Change-Id: I9b89f169d96959daa0fbf961e9bd6019b21c98b0