[roll] Roll topaz [cleanup] remove deprecated lib.settings from topaz
This CL removes the deprecated settings and setui dependencies.

Deprecated settings code is being migrated to experiences so it can be removed
from the topaz repo (and eventually removed from experiences once new settings
feature parity is met).

Original-Bug: 45550

This is 3 of 4 CLs that need to be submitted in order:
1. [fxr/381099] add SDK visibility to experiences/settings
2. [fxr/380992] copy lib.settings into experiences
3. => [fxr/381113] delete lib.settings from topaz
4. [fxr/381106] remove SDK visibility from topaz

The set of CLs has been vaildated with:

        fx make-integration-patch

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/topaz/+/381113
GitOrigin-RevId: 6396c204ca92fc3e58f7dbf5b4867aab3d912ee9
Change-Id: I7e97c8d651307c8c063cbf386b326db5f308ee91