[roll] Roll fuchsia [fidl][measure-tape] Generalize expression code
Instead of generating C++ expressions, we represent expressions in an
AST and later transform these into valid C++. For instance, the
expression

    FIDL_ALIGN(value.length() * 16)

is represented as the AST

    FidlAlign
    - Mult
      - Length
        - Local of "value"
      - Num of 16

This paves the way to a direct translation to Rust, while reusing the
biggest part of the code which emits the expressions and code blocks.

This representation also opens the door to simple constant folding, and
combining statments.

Test: fx test measure-tape-tests
Original-Bug: 50011
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381753
GitOrigin-RevId: 08abed28a42ce9f3ca3eb47e938d7d3da960b889
Change-Id: Ia4e8a24e83a5f3e1186bfc7653d844318665960c