[roll] Roll fuchsia [logging] Make fxl_logging a source_set
fxl_logging has always been a source_set on host. However,
on Fuchsia was defined as a shared_library because it contains
global state. With the recent migration of the FXL library to
use syslog, we can just rely on the global state in the syslog
shared library and not keep any state in FXL.

This CL introduces separate logging backends for host and
Fuchsia to reduce the number of ifdefs in the code. There is
opportunity to encapsulate more functionality in the backend
interface but this CL only introduces functions related to
modifying the global settings.

fxl::GetLogSettings() has been removed because we can't get the
log file path from the syslog library. The few clients that
used this function have been converted to not depend on it.

Original-Bug: 40982
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384221
GitOrigin-RevId: ce3bd7df69841bac552a3d6e94926df2dc3e6d67
Change-Id: Ic2ef3bf896ce19de144ea43230505f5ec9014c31