[roll] Roll fuchsia [brcmfmac] Fix alignment for sim IovarSet()
The sim implmenetation of IovarSet() relies on being able to cast
directly into the buffer provided, which may be misaligned for the
desired access.  Fix this by making a suitably aligned copy of the
buffer before passing it into IovarSet().  While this is another copy
step, it should not be significant for a testing framework such as sim.

Original-Bug: 45759
Test: fx run-test brcmfmac_sim_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383904
GitOrigin-RevId: a0c91ecde77097d6e7678b8480fd3af1013f22a1
Change-Id: Id7313c1c51bf071bf25f0d4aa6498af22bccd5cd