[roll] Roll fuchsia [gpio] Configure interrupts before creating them
This change ensures that the configuration of GPIO interrupts occurs
before the interrupts are created, so that the configuration change does
not cause a spurious request on the new interrupt.

This change also fixes a shift overflow on the interrupt select
register, that due to undefined behavior was returning correct results
only accidentally.

Tested: Booted Astro device and created GPIO interrupts on Astro device
    using brcmfmac driver
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/370695
GitOrigin-RevId: 6ffaac3ae5388942a0ad714b78a5d8d1bf87ed5e
Change-Id: I8ee9243a94c250bad3e050aca8b14981d95ba339