[roll] Roll fuchsia [docs] Rewrite and improve sysmgr-configuration.md
This updates sysmgr's primary documentation, adding a lot of missing
detail about the configuration format and in general document more of
the pitfalls to avoid when modifying sysmgr configuration. A new section
was also added to document how to add to sysmgr's configuration and the
recommended best practices for avoid sysmgr config conflicts.

Original-Bug: 48223
Original-Bug: 48215
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/374081
GitOrigin-RevId: f37f3ae5506e1381768bf9f1dea4bb944eeae9bf
Change-Id: I2ccc0f92f03ee28218be5abfa49e29457e088d1c