[roll] Roll fuchsia [brcmfmac_test] Fix timer_test flakeness
The flake in timer_test is caused by the async mechanism of timer,
originally the call_timerset and call_timerstop bools are set after the
timer start, so we don't know how many times a periodic timer will fire
before these values are actually set, the test cases in which timer_cnt
is checked will fail because of this. Now these values are set before
the timer start, the issue should be fixed.

Test: manually run timer_test for 100 times.
Original-Bug: 50191
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384639
GitOrigin-RevId: 69087e8fa51733615329724db35318456afbb85a
Change-Id: Ic7a1d22779921b3f8c939a1601a8cb75f0d97b7a