[roll] Roll fuchsia [mediaplayer] fix test flake
mediaplayer end-to-end tests are necessarily sensitive to real time,
because we don't have a distributed fake clock. The test bot
environment is such that the fake audio renderer can get behind to
the point that the VMO used for audio packets fills up and packets
are dropped. Verifying the contents of the packets is therefore
flakey.

This CL removes the check that verifies that actual audio packets in
the offending PlayBear test.

TEST: Change to test

Original-Fixed: 50016
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384684
GitOrigin-RevId: f113d4039a5e3756b3ae7a2c31b5d1c3092b7c44
Change-Id: I654b5b7a38d95e0b4cf81f3047eac0846e5fc1eb