[roll] Roll fuchsia [zxtest] Allow comparing chars in C
Prior to this, the _Generic switch used to dispatch included cases for
various sized integer types, floating point types, strings (const
char*), and other pointer types.

char, unlike signed char and unsigned char, is not compatible with any
of those types. This sense of "compatible with" is from 6.2.7 of the
C11 standard.

signed char is compatible with int8_t already in the case list, and
unsigned char with uint8_t.

This includes tests for all three types, just to make sure.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384814
GitOrigin-RevId: 3e4b7eb2579eb01be77fd21426e9f0651b0a68b5
Change-Id: If8de8798076bd52a5733a7bd34d9d88dcc5d3a6d