[roll] Roll fuchsia [garnet_input_latency_benchmarks] Switch to Dart-trace-model version of test
Switch to using the results from the newer version of this test that
uses the Dart trace model.  (i.e. Use those results for perfcompare
and for uploading to Catapult.)

Remove the SL4F wrapper for the older version of this test.  (However,
note that the older version of the test still gets run by the non-SL4F
test runner in src/tests/benchmarks/benchmarks.cc.)

Original-Bug: 23043
Test: on CQ: terminal-x64-release
 * can be run locally with:
     fx set terminal.x64
     fx run-e2e-tests garnet_input_latency_benchmarks_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384563
GitOrigin-RevId: 697742ff03595203d5b82659032a67c3a72a479a
Change-Id: Ic45867dab1eac1e4b12e7fd3ea70c96ff4e09453