[roll] Roll fuchsia [fidlc] Recoverable compilation
Don't return early from Library::Consume and Library::Compile methods
when an error is encountered; get a checkpoint in the ErrorReporter at
the start of the method and return false (signal failure) only at the
end of the method if there were errors.

This way we can print out more than one error in the consumption or
compilation steps.

Test: fidlc host tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383206
GitOrigin-RevId: 1805b8e5a703cbb7ab949e9009c351ce351dfe61
Change-Id: Ie5cced5a1ca502c9b7afaa11e343cdeea8d3fd75