[roll] Roll fuchsia [debugger] Enable hit_count display for breakpoints
hit_count is already collected and synced between debug_agent and zxdb.
This patch enables the display of such number when listing breakpoints
using "bp".

The behavior of "bp" is changed. A new column called "hit-count" is
added. This column will be empty for pending breakpoints.

The hit_count of a breakpoint is also available as a setting and can be
obtained by "bp get hit-count". Currently there's no support to set this
value.

Original-Bug: 43929
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384296
GitOrigin-RevId: f37d8b7bf053f20cedb1737a4a94d58dc50ea397
Change-Id: I7c5cd4e8a391c4d81dee951569f5413069364145