[roll] Roll fuchsia [utest][fs] Remove race in mmap_crash tests
After performing munmap there is no guarantee that any operation (from
any thread) that performs a mapping or creates a VMAR will not reuse
the portion that was freed. As ASSERT_DEATH creates a thread, which
requires creating several memory mappings, it is possible for the
mapping to be reused between the munmap and the access.

This change moves the munmap into the death callback to remove any
mapping calls from happening before the access test can happen.

Original-Bug: 41601

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384774
GitOrigin-RevId: c433f446c8adf121ec94ea9bda637e5d3d0b4e55
Change-Id: I8e60e9628f5c15142bb668e74ea54142678f52e9