[roll] Roll fuchsia [kernel][heap] Fix cmpct_init() bug
In the introduction of the cached OS allocation we forgot to null that
out in cmpct_init(). This doesn't produce bugs in production, as there
is only ever a single instance of cmpctmalloc's backing heap object;
however, this does now produce bugs in tests where we end up
unintentionally reusing the cached allocation across test cases.

Original-Bug: 49123
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384729
GitOrigin-RevId: 215b4e105f96b9d058daa47d74deb86cc7f191de
Change-Id: Ie4a711a6798932e382d663e9fae631aef9cc1b21