[roll] Roll fuchsia [kernel][heap] Update cmpct_get_info()
This change
(1) renames "size_bytes" to the less ambiguous "used_bytes";
heap_get_info()'s size_bytes also is updated to output the "total bytes"
(which "used" + "cached", which is a more informative quantity at that
level.
(2) extends the signature to also fill in the value for the size of the
cached allocation
(3) permits null parameters (which are ignored) for convenience.

Original-Bug: 49123
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383597
GitOrigin-RevId: 10d316330afeeced44a101a928c2c7d9d01a562c
Change-Id: I650af273841ec5c64f8fcd6b24b00567708964f0