[roll] Roll fuchsia [fbl][wavl] Remove a useless and incorrect using directive.
There is no real reason for the generic node_state fetcher in WAVL
tree to have a RankType using alias.  In addition, this alias was
incorrect.  It was missing the TagType parameter for the node_state
template which caused it to incorrectly fall back on the default tag.
This would only be caught when attempting to instantiate a tree of
objects which used the ContainableBaseClasses pattern, but no default
WAVLTreeContainable mix-ins.

Also, add new tests which would have caught this in the first place,
and fix an existing test which was broken and caught by the new test
augmentations.

Test: new tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383486
GitOrigin-RevId: a8f281a9da6c613990851e89d9ba4deca9fe78c8
Change-Id: I838f7ec0ead8a95b10bfcd5b85af9f68caf18075