[roll] Roll fuchsia [netemul] Moves gRPC client stress test to flaky.
The stress test repeatedly sends small messages to a guest
VM.  This test was written to demonstrate testability of a
change that fixed a race condition in this communication
mechanism.  In order to attempt to excite the issue, 1000
iterations of the gRPC operation are performed serially.
This test was written before the 5 minute timeout was
imposed on individual test runtime.  It seems that the test
has been exceeding the timeout a couple times per week.
This change moves the sometimes long-running test to flaky.

Original-Bug: 49752
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384637
GitOrigin-RevId: 9f7779d0fb28ff6730f8b38805a166488da7bc02
Change-Id: I343ad968bf844b46ae3af31775feeb9b2b3ad169