[roll] Roll fuchsia [kernel] Fix panic caller address.
The current definition of "panic" is a macro which expands to a call to
a real function "_panic" as follows:

  int my_buggy_function() {
     ...
     _panic(__GET_CALLER(), __GET_FRAME(), "panic string", ...);
     ...
  }

Unfortunately, this results in panic printing out the program counter of
the function that _called_ "my_buggy_function", and not the panicking
address inside "my_buggy_function" itself.

This CL moves makes calls to panic real calls, and then makes the panic
function itself call "__GET_CALLER()", resulting in the PC of the call
site to panic. We additionally update the frame pointer to point to the
frame of panic. The full stack dump is shown below the panic message.

As a nice side-effect, the reduced number of arguments to panic (and
thus simplified code) removes ~40kiB of code on x64 release builds:

                         text      data     bss      total
     ----------------- --------- -------- -------- ---------
      base              1683892   504896   594763   2783551
      simplify panic    1642484   504928   594747   2742159

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384035
GitOrigin-RevId: 60169de761f6cbd44716ff0976924969d26af670
Change-Id: I61a58873fe3a57f5c973f8e111c3b9de1341d3b9