[roll] Roll fuchsia [fidlc] Improve test coverage of errors
Currently, 29 of the 105 errors and warnings in fidlc are not covered
by any of the fidlc host tests. This CL adds tests for every error that
is not covered to prevent regressions in future changes to fidlc.

...with the following exceptions:
* ErrGeneratedZeroValueOrdinal: to trigger this error, we would have to
  somehow hash a method's library + protocol + method name to 0 --
  which we had better hope is not possible to do deterministically.
* ErrCouldNotResolveIdentifierToType: this seems to require something
  to go wrong internally in fidlc, since this error is triggered by the
  result of a binary operation being assigned to a constant of an
  unresolvable type -- and the type would have already been caught as
  an ErrUnknownType. (Maybe it should be an assert and not an error?)

This change doesn't give us perfect coverage of course since most
errors can occur in different contexts and from different causes, but
some signal is better than no signal.

Test: fidlc host tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383477
GitOrigin-RevId: a7c467593aed440f52923a09f3778e5525a7a3a5
Change-Id: I23cf93e4812f4dbd95fc3237ebe12583b6d2a9f8