[roll] Roll fuchsia [wlanstack] Mark active client iface
We do not remove old iface node from Inspect tree right away. Thus, if
the device iface has been destroyed and recreated, it would be difficult
for clients to determine which iface node to retrieve information from.
In this CL, we add a property in the wlanstack Inspect tree to indicate
what the active client iface node is.

Original-Bug: 48950
Test:
1. Added unit test.
2. Verified Inspect log while manually destroying/creating iface.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384304
GitOrigin-RevId: cf4007400e74b95215027b8baa02971ac53fa37e
Change-Id: Ic5f3574e36b14e66af26069f8b7aa966db5d701c