[roll] Roll fuchsia [kernel] Remove unused "target_set_debug_led" method.
"target_set_debug_led" allows targets to implement methods for switching
a debug LED on and off. Currently, we only call it in the scheduler (to
indicate if the system is idle), but don't actually implement it for any
target or platform.

Remove it for now, and add it back later if the need arises.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383973
GitOrigin-RevId: 8e119a8cfedaa571d2ddec837720cb64a96b7497
Change-Id: I24c2407a830f18c90bfa03fcc452f52b4ba57ace