[roll] Roll fuchsia [kernel] Adjust kernel init levels.
In particular, add a "{arch,platform}_prevm_init" functions that are
called before virtual memory is (fully) enabled, but after the command
line and UART is available.

This is to facilitate a non-trivial chunk of code that runs on x86 which
needs access to the command line but also needs to run prior to
code-patching.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383692
GitOrigin-RevId: ece30ffdd77b58b6d171a9b0a330f9dba1c7a722
Change-Id: I251b66751fb693cf79ad17ba46baf30048499f76