[roll] Roll fuchsia [bt][sm] Fix MTU handling in SMP refactor
Handle minimum SMP MTU validation by asserting during SMP initialization
instead of validating packet length is < the MTU upon packet reception.
We already validate that packet lengths match the payload associated
with their SMP header code, so validating that they are smaller than the
spec-prescribed channel MTUs is unnecessary.

The important check is that the SMP fixed channel's MTU is large enough
to support SMP, which is safe to ASSERT on for a spec-compliant L2CAP
implementation. We check whether the MTU of the SMP fixed channel is >=
65 bytes to determine whether the local device supports Secure
Connections (see spec V5.0, Vol. 3, Part H 3.2). This is used in Phase 1
to construct the AuthReq field.

Also add an MTU parameter to FakeChannelTest::ChannelOptions, so MTUs can
be tested easily in a FakeChannelTest.

Original-Bug: 49077, 46278
Test: `fx test bt-host-unittests -- --gtest_filter="SMP*"`
      Removed MTU length validation tests from sm/packet_unittest.cc.
      Added MTU validation to pairing_channel_unittest.cc. Modified
      phase_1_unittests.cc to use the MTU size to signal whether to use
      Secure Connections.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381453
GitOrigin-RevId: 4076b2b587a911dd4e8936ed6b2eb686665f1bb5
Change-Id: Ib3d404f3c4877d7fbd1b3040373d4fe31b7f164d