[roll] Roll fuchsia [isolated-ota] launch the package resolver
This CL adds support for launching the package resolver
and package cache inside a sandboxed environment, against
the pkgfs already launched and using a provided repo configuration.

We use the integration test variants of pkg-cache and pkg-resolver so
that we can override paths like /config/data with the configuration
that's been provided. To ensure that these are available, we add an
'isolated-swd' package that the 'isolated-ota' library depends on that
contains all the relevant binaries and manifest files.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377857
GitOrigin-RevId: 8d808a289c09e411f1e645f79a358a24963244b7
Change-Id: I3e2bae3d0f5e69fde8e24313e9e0dd86b81d1da9