[roll] Roll fuchsia [flatland] SetImageOnTransform implementation.
Similar to SetLinkOnTransform(), attaching an Image to a Transform is
simply making its representative TransformHandle the priority child
of whatever Transform the client attaches it to.

Since there are now two types of content, setting one on a Transform
removes any other content that was on that Transform.

At the moment, ComputeGlobalImageData() computes the ordered list of
ImageMetadatas for rendering. In the future, this function will most
likely output more, such as the matrices associated with each Image
or data for group opacity.

TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383322
GitOrigin-RevId: 4c276d0354302b3a898b92fab17c568229f6f095
Change-Id: I4153f00aec3de95442856e594b23f2f7bda69b4f