[roll] Roll fuchsia [intl] Factors intl_strings from strings_to_fidl
This is a good point at which reusable bits of the strings_to_fidl
code can be factored out into a stand-alone library.

The bulk of this commit consists of moving the code around.  However,
some changes were needed such as adaptation of the BUILD.gn files and
README, removing unneeded bits from both the binary and the library end,
and factoring out the mods message_ids and parser to stand-alone files.

Original-Bug: 50722
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384753
GitOrigin-RevId: 01096ee87ed46804a14482de4e8cf3e4cf149526
Change-Id: I649825dfc9dd59a85d46ed57168e3e4aeea6da87