[roll] Roll fuchsia [blobfs] Make write-compression configurable
Adds a MountOption and a startup flag (--compression [algo]) to
configure the algorithm blobfs should use to install new blobs.

This is mainly useful for integration tests, to allow them to exercise a
specific type of compression algorithm.

The old flag (--write-uncompressed) is still supported to make this a
soft migration (just so we can break up this change from the changes to
the generic fs code to plumb the flag). Setting both flags has
indeterminate behavior.

Tested: Compiled, booted, ran tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384223
GitOrigin-RevId: 33f56f6b8afa0bea1bd436baaee0595725b1ab23
Change-Id: I241bed702ae01e214c1f14afad054575ad57ee7b