[roll] Roll fuchsia [camera] Serialize stream start up (workaround)
It should be legal to start up all streams (in a single config)
in parallel without race conditions or other conflicts. However,
in the current state of the driver, the streams must be started
in sequence, and fully serialized. No part of one stream's start
sequence may overlap with that of another.

This CL will work around the limitations. fxb/51218 will track
the restoration to normalcy.

Original-Bug: 51219,42241,51166
Test: Run 10x with no start up issues
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385213
GitOrigin-RevId: 72d046a0334ac0ff6ecfb1c8944c57250068ef91
Change-Id: Id93a54c5f3cf84216ff4efb6685b11f686903488