[roll] Roll fuchsia [network-mgr] Adds configuration for qemu
Due to the way that topological paths currently work, when running
under qemu the emulated topological path needs to also be configured
so that network manager can configure the emulated interface correctly.

This change introduces a `default_interface` that can be used to
configure these unconfigured devices.

Also, this change narrows the tests that the router QEMU buildbot runs
because, for example, are not including the bluetooth or lowpan stacks.

Original-Bug: 50780
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383881
GitOrigin-RevId: 2ffbff4665827a770d1a2635ec86a37316e31b54
Change-Id: Icba6e1a0f6e2176eed771bfab33fa3459cfc97c2