[roll] Roll fuchsia [intl] Modifies message ID generation
Generates the message ID based on the message name and content.  This
is needed so that messages which are homoglyphs (written the same) in
English, but not in other languages can coexist:

Example:

```
// <!-- читати -->
// <string name="action_read">Read</string>
// <!-- прочитане -->
// <string name="label_read_messages>Read</string>
```

While this is not ideal (we probably want to mix in other stable
content), it is enough for our purpose as long as both the message IDs
and the message bundles are generated at the same time during
compilation.  This works because we currently don't mix in messages from
libraries, nor allow building the messages independently from the code.

Once that changes we will need a more enduring message ID scheme. The
good part is that we can change the scheme at any point before that
happens.

Original-Bug: 50722
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385378
GitOrigin-RevId: f4cdc40bca1e340d097cf457d05a7cdc084df787
Change-Id: Ibccc697b48e6c2a31b7cdead3a9b204d4ec3ccaa