[roll] Roll fuchsia [archivist] Report "ERROR:" klogs as Severity::Error
In order to get a better read as to the number of error logs
produced by the system, we need to count klogs as well. Prior to
this CL, all klogs were reported as Severity::Info. This CL is a short
term workaround, until we have structured klogs, to look for the string
"ERROR:" and "WARNING:" to map to Severity::Error, and Severity::Warn
respectively.

Original-Bug: 51179
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384900
GitOrigin-RevId: 9c64763237d8f269aa4b4e2218fdf2970d0aeea6
Change-Id: Ie467988e0d56d665d7da1f6aa7eb82dbf912f410