[roll] Roll fuchsia [modular] Re-land "[sessionmgr] Replace ledger with assorted std::map"
A bug in the previous iteration of this change would notify clients of
StoryProviderWatcher of a story being deleted even in the case that the
story was already deleted.

It turned out one of our clients was relying on this behavior, but there
was no test coverage :\. This CL is landing after a change which adds
the required test coverage.

Original-Bug: 49677
Test: fx test session_shell_test; Also manually verified the test fails without the fix.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385000
GitOrigin-RevId: 22473fbd1d5858b5de7d8b5f37e60913bf3f5dc2
Change-Id: If83f20af056eef84146b38b40542306ead241645