[roll] Roll external/github.com/catapult-project/catapult Cleanup of how Job generates issue tracker updates.
The motivation for this is to make it easier to improve the issue
tracker updates posted by Pinpoint, but significantly changing the
content of those updates is not in scope for this change.

Mostly this change is a refactoring:

* extracting a bunch of logic for generating
issue tracker updates about found differences from job.py into a
new module.
* keeping state in simple explicit objects, rather than having many
variables and function parameters and return values with implicit
invariants between them.
* rename various functions to have clearer names, and add many more
comments.

However there is a small opportunistic bug fix due to simplifying
the logic in _ComputePostOwnerSheriffCCList (now called
_PeopleToNotify), so this adds a unit test for that.

Original-Bug: chromium:1076756
Original-Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2172930
GitOrigin-RevId: 5a865628d10a6fd89547a7ad0828984c2472e447
Change-Id: I0095c9c6e0b323d7fdff1b18a8f1603c304d5456