[roll] Roll topaz [fidlgen_dart] Exclude declarations based on BindingsDenylist
This CL adds support for BindingsDenylist in fidlgen_dart, completing
the work started in I56c9ae065381b1f5e2ba4cb7c7286c9f7ae1b4af which
added support in all other fidlgen backends.

The attribute is used like this:

    [BindingsDenyList = "dart"]
    struct Foo {};

In this example, the struct Foo would be absent from the Dart bindings,
as if it was never written.

Test: goldens
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/topaz/+/385500
GitOrigin-RevId: c8fc056fc4cb3dbe9af907d3539b054d5d2fc0af
Change-Id: If7f66d7200b56dc9acc573dda0ec801e5f8ca3be