[roll] Roll fuchsia [kernel][x86] Add unit test checking Spectre V2 and MDS hot patches work
lib/code_patching is used to dynamically select the correct retpoline
thunk and the MDS mitigation appropriate for the current processor and
the mitigation cmdline settings.

Two bugs have recently meant that mitigations settings were locked in
before the command line was parsed; this meant the executing text was
not the correct one for the current processor + settings. Add a unit
test that checks whether the running speculation mitigations are the
correct ones.

Original-Bug: 33667 Spectre mitigations?
Original-Bug: 45628 improve spec mitigations control bits

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385280
GitOrigin-RevId: 443ff6057852ce1219f3959af679375a4cd060e7
Change-Id: I497f2525a081a7004db88198083953583cdcf8d5