[roll] Roll fuchsia [fs] Make blobfs compression configurable.
Adds a new build argument 'blobfs.write-compression-algorithm' which
takes a string like 'ZSTD' to control blobfs' default compression
algorithm.

The values are defined by blobfs, and the following are supported:
- ZSTD
- ZSTD_SEEKABLE
- ZSTD_CHUNKED
- UNCOMPRESSED

If no value is specified, the system default is used.

The legacy 'blobfs.uncompressed' flag is transparently converted to the
new flag, to make this a soft migration. The replacement value is
'blobfs.write-compression-algorithm=UNCOMPRESSED'.

Tested: Built with `blobfs.uncompressed` set, verified 'UNCOMPRESSED'.
Tested: Built with new flag set, verified the value is passed through.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384895
GitOrigin-RevId: a333581bae794d2761cff95de2b9db0a36b327cb
Change-Id: If8e5e70a5c9274ba3c9153590b375c90445b8a80