[roll] Roll fuchsia [camera][scenic] Conditionally set sysmem buffer names.
This change makes the camera stack and scenic ImagePipe2 less aggressive
in setting the debug name of objects. Instead of unconditionally setting
the names of negotiated sysmem buffers, they use the MaybeSetObjectName
function to set the name only if it is still the sysmem-default name.
Additionally, this change alters where scenic sets the name of buffers
passed to ImagePipe2, setting them only when a specific buffer is added
via AddImage instead of when the collection is initially added.

Test: Used memgraph tool to verify expected debug object names.
      Extended DeviceTest.RequestStreamFromController to check vmo names
        provided by camera.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385976
GitOrigin-RevId: 6e8fa72d8a1ca5a4bf36e0695c917bd32fc80005
Change-Id: I4e7ce4c3fd4e2515306bcca607677e3e78b4b609