[roll] Roll fuchsia [fbl][string_piece] Disallow constructing a fbl::StringPiece from nullptr
Eventually, we will migrate all users of fbl::StringPiece to using
std::string_view directly. The latter does not allow construction from
a nullptr.

By marking the constructor = delete in this CL, we ensure we have
found all call sites. A later CL can remove the declaration if needed,
before the entire class disappears.

As a small drive by, a call site's comment is tweaked to be more
accurate.

Original-Bug: 33159
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385697
GitOrigin-RevId: f33a3892733a96f3cbb4be97687e37cd970242ba
Change-Id: Icdbd127f53f57a97f3f108f0d631421e8dbcdfd3