[roll] Roll fuchsia [GN SDK][CQ Tests] Run GN SDK script tests in CQ
This adds the tests in //scripts to the test bundle and the
buildbot:config.

These are unit tests for the f* scripts which are shared with
teams that are building components in GN based projects.

Multiply: gn_sdk_script_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/358229
Original-Revision: 0907e08ff06974c112aee99c4c52cb07657bde6d
GitOrigin-RevId: ae016722edd3f477461950a2368a2d40eeda0fbc
Change-Id: I17909a49bebf00f44d63a76655d0d05f729b7d5c