[roll] Roll fuchsia [fbl][containers] Add AllowMultiContainerUptr NodeOption
Add a new NodeOption flag that users may set if they wish to use the
ContainableBaseClasses helper, but also want to allow their objects to
be tracked in one or more containers using unique pointers.

See the comment for the option in the code, as well as the tests, but
the TL;DR here is that while attempting to do this would _normally_
indicate a mistake, and should not be allowed at compile time, there
are actually legit situations where someone might need to take this
safety off.

Also, make the check in ContainableBaseClasses a bit more strict.
Prior to this, CBC would allow a unique_ptr in the list of
containable, but only if it was the very last containable in the list.
While this implies uniqueness of the unique_ptr, it felt like an
unreasonable requirement that users know that the behavior was
permitted, but only if the containable mix-ins were listed in a very
specific order.  Additionally, this still would not allow the user to
have multiple contianer nodes which used unique_ptr which were meant
to be used in a disjoint fashion.

Instead, always complain if a user attempts to use unique_ptr along
with ContainableBaseClasses, but point them at the option when we do
so, so that the user can read up on the best practice and make the
right call for their specific situation.

Tests: new unit tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384293
Original-Revision: 15a80185a054207ffd5204d3d2aba1327c5f1c1e
GitOrigin-RevId: 7f8f5ede3826d3c8a0c274bca051612d764a132a
Change-Id: I67657862fbafdcfa56f2d92b9cbd75c4a4cdf5fb