[roll] Roll fuchsia [gidl] Test recursive unions in GIDL
This CL adds GIDL conformance tests for unions that recursively contain
themselves (1) in a vector and (2) via mutual recursion with a struct.

The tests are disabled in Go because using these types causes a stack
overflow (fxb/38191), and in Dart because the Dart analyzer complains
about self-references (fxb/7644). In the case of Dart, we also have to
hide the FIDL declarations using BindingsDenylist, because the analyzer
problem happens when compiling the bindings.

Test: run all GIDL tests
Original-Fixed: 36424
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386198
Original-Revision: 50be52828e03baa3f40862c24cca492d0579ac0d
GitOrigin-RevId: e4f1fef8c11113b8832688595f016be115d2c258
Change-Id: I031cf7afd4b39e7cc35bd24f0c28c1f163b59151