[roll] Roll fuchsia [SetUI][core] Remove startup result receiver.
A result receiver was previously handed back with the
environment on spawn in order to allow spawn to be
synchronous, but a caller (mostly tests) to monitor
the end result.

Recent changes have modified this code path to
allow for spawn be awaited on by executing on a
single thread. Doing so allows for waiting on the
environment to bring up before returning.

This change makes all environment spawns
awaited on and removes the callback receiver.

Original-Fixed: 50378
Test: fx test -o setui_service_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381434
Original-Revision: c60ed1d5b1de692406ac8b34d2ddf89153ff8b3a
GitOrigin-RevId: 400babcd17654370f556a4f92b9893b7d5539eac
Change-Id: Ibdb2c878145f96a3ffc8ebd55c5c4f0d3274d1b4