[roll] Roll fuchsia [sdk] Properly handle the deps attribute of sdk_source_set.
Previously this attribute was used to attach dependencies that would not
be reflected in SDKs. This means that public_deps had to be used for these
even for dependencies only used in the implementation of the source set.

Instead, deps are now added to SDK deps, and a new non_sdk_deps should be
used to attached non-SDK deps (e.g. codegen actions).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386739
Original-Revision: 0f559315155ea84cf1a3fe22b39a736cc7e6a8a2
GitOrigin-RevId: 93153d7385c8d17684b7361d6dd97888595f7309
Change-Id: I5dd306198c355995eae2b1cac81648bcd3876fb3