[roll] Roll fuchsia [src] Delete //src/ledger from fuchsia.git
Now that modular has replaced all ledger usage with std::map, the ledger
code itself is unused and no longer maintained in fuchsia.git, so it
creates no value sitting here in duplicate.  This change removes it in
its entirety.

This change also makes a first-pass attempt to remove other references
to ledger from the build, testsuites, and benchmarks.

There are a few places in the docs where ledger is still referenced but
those docs are notably stale in other ways and I believe rewriting those
documents is out of scope for this change to remove the vast majority of
references and source.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386422
Original-Revision: bb21a62d8a9358afab3424e3b300ee5185a768a5
GitOrigin-RevId: 82cb59160f3c100d3f01a745546a42db74ad25e3
Change-Id: Id8f179e383183b252025cd84182a720b7a8b1714