[roll] Roll fuchsia [run_test_component] Add config to allow Error logs
Adding config to allow error logs for certain test urls.

The feature to fail on ERROR is not active yet and this is first step in
implementing that feature.

The config is read from base package, so developers need to do fx update
when adding a test url to the config.

BUG=49735

TEST=added unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386433
Original-Revision: 87ac8008e8c75605981414b0433d5caf85feafe4
GitOrigin-RevId: 8a38120e3ee26da9744643e9842bdfe1f8f40f90
Change-Id: I26aca82d2da73fdbab8c81afd1ad3f9874fff38a