[roll] Roll fuchsia [kernel][wait] Encapsulate the data structure inside wait_queue_t
We'd like to eventually change this data structure. By giving it
names, we are taking a step to separating the data structure (i.e.,
how the collection of threads is laid out, adding to or removing from
the collection, finding a thread in the collection) from the wait
queue operations (waking a thread, etc.).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385232
Original-Revision: 634f9f920f3086ad4c1530d69ea6b57dea94beea
GitOrigin-RevId: 896d090dd372c65fd07b5929bc450bd47620a2ad
Change-Id: I3c236ee92c792668cf1529924e3c8a8c3086a640