[roll] Roll fuchsia [paver] fix sherlock bootloader read and size
We had previously fixed SherlockBootloaderPartitionClient to properly
skip the block0 metadata when writing, but reads were still starting at
block0, with the effect that we would pave the bootloader every time
since the check for content equality always showed a mismatch.

With this CL, the paver properly skips writing the bootloader if it's
the same as what's already in the partition.

This also fixes the reported partition size to omit the first block,
since it's not really writable.

Test: `fx pave` and `fx ota` correctly skip matching bootloader image
Original-Bug: 50999
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384914
Original-Revision: d36370bc637ae733ef94fd51d72b6ff766f79ae3
GitOrigin-RevId: 380d76deeeb5ee863b53f986f1fed0c46481dd87
Change-Id: I1a7a2f53565bc206876e51d1bb5ed5ccc46c072e