[roll] Roll fuchsia [region-alloc] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
allow regions to exist in two trees at once.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386815
Original-Revision: c717a32a1b16b3f135418359c1c0639733bf8e8d
GitOrigin-RevId: 1eb65f9eebd51be43613d326f4476bf321b3a431
Change-Id: Ie54efecfe6d33fba7a0bd381625a51bcb116d41d