[roll] Roll fuchsia [rust] Add hypen to extra-filename
Based on how cargo does this, and how the -o filename is set, and how
the dep-info is set, I am pretty certain that we want filenames to be
like: libfoobar-123deadbeef.rlib instead of libfoobar123deadbeef.rlib

It looks like gnaw missed a hyphen, so this adds it back, fixes the
tests and updates the third_party build file. This wouldn't be visible
in practice because we specify the output filename explicitly, but it
shows up if you try to reuse the generated command for other things.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385253
Original-Revision: c87e78de3090b0a1a9bbf0017d62c49c1040a481
GitOrigin-RevId: 81e604ad33453dd440c6e45726b5c87ac3de90c8
Change-Id: I2f42b3add83caf9a5e99b3123ba606fe7c29f3e2