[roll] Roll fuchsia [kernel][thread] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
implement a ProcessDispatcher's list of ThreadDispatchers.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Tests: existing unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386752
Original-Revision: 899638d5321fe85711d49e4ea6733d65716be84f
GitOrigin-RevId: 5fc16ebfd2d41f83babdec63dbc94a2b31d3e587
Change-Id: I2b716c0458b0eeb6a513b6ab6eeec87eaa7970aa